gui/requirements.txt
PyQt5
pyserial
matplotlib